+++ src/intrpr/cfg_mgr.py
@@ -10,6 +10,7 @@ import utils.err_codes as uerr
 class Cfg(ty.NamedTuple):
     prompt: str
     pth: tuple[str]
+    aliases: dict[str, str]
 
 
 def sandbox_runpy(fl: str) -> dict[str, ty.Any] | None:
@@ -30,6 +31,16 @@ def sandbox_runpy(fl: str) -> dict[str, ty.Any] | None:
         return None
 
 
+def err_typ_repo(key: str, expd_typs: tuple[type], got_typ: type) -> None:
+    comma_sepd = ", ".join(f"'{i.__name__}'" for i in expd_typs[: -1])
+    or_sepd = f"{' or ' if comma_sepd else ''}'{expd_typs[-1].__name__}'"
+    expd_str = comma_sepd + or_sepd
+    ugen.err_Q(
+        f"(config) Expected {expd_str}, got '{got_typ.__name__}' for key '{key}'"
+    )
+    ugen.info_Q(f"(config) Using default value for '{key}'")
+
+
 def get_cfg() -> Cfg:
     """
     TODO: Fill up!
@@ -41,9 +52,7 @@ def get_cfg() -> Cfg:
     """
     prompt = uconst.Defaults.PROMPT
     pth = uconst.Defaults.PTH
-    err_report = lambda key: ugen.err_Q(
-        f"Invalid type for config option '{key}'; using default value"
-    )
+    aliases = {}
 
     cfg = sandbox_runpy(uconst.CFG_FL)
     if cfg is None:
@@ -51,22 +60,44 @@ def get_cfg() -> Cfg:
 
     for key in cfg:
         val = cfg[key]
+        # To continue outer loop when checking iterable values
+        valid_val = True
 
-        # Only callables and strings are allowed for prompts
+        # Callables and strings for prompt variable
         if key == "PROMPT":
             if not (isinstance(val, str) or callable(val)):
-                err_report(key)
+                err_typ_repo(key, expd_typs=(str, callable), got_typ=type(val))
                 continue
             prompt = val
-        # Only a tuple of strings are allowed for the path variable
+
+        # Tuple of strings for PATH variable
         elif key == "PTH":
             if not isinstance(val, tuple):
-                err_report(key)
+                err_typ_repo(key, expd_typs=(tuple,), got_typ=type(val))
                 continue
             for i in val:
                 if not isinstance(i, str):
-                    err_report(key)
-                    continue
+                    ugen.err_Q("Expected 'str' values in value of '{key}'")
+                    valid_val = False
+                    break
+            if not valid_val:
+                continue
             pth = val
 
-    return Cfg(prompt=prompt, pth=pth)
+        # Dict of string: string pairs for aliases
+        elif key == "ALIASES":
+            if not isinstance(val, dict):
+                err_typ_repo(key, expd_typs=(dict,), got_typ=type(val))
+                continue
+            for k, v in val.items():
+                if not isinstance(k, str) or not isinstance(v, str):
+                    ugen.err_Q(f"Expected 'str: str' values in value of {key}")
+                    valid_val = False
+                    break
+            if not valid_val:
+                continue
+            aliases = val
+
+    return Cfg(prompt=prompt, pth=pth, aliases=aliases)
+
+